pyproject.toml
[tool.poetry]
name = "theinnerpalace"
version = "0.1.0"
description = ""
authors = ["Nicolò Pignatelli <pignatelli.nicolo@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
pyautogen = "^0.2.27"
packaging = "^24.0"
logging = "^0.4.9.6"
colorama = "^0.4.6"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
websockets = "^12.0"
pypubsub = "^4.0.3"
pydantic = "^2.7.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
